refactored to remove circular dependency issue
